+++ qmm/fileutils.py
@@ -17,11 +17,24 @@ from qmm import bucket
 #  in order to build paths corresponding the the game expectation.
 # NOTE: race folder contains it's how set of sub-folders. Might need to be
 first_level_dir = (
-    "items", "outfits", "setBonuses", "statusEffects", "race", "colours", "combatMove"
+    "items",
+    "outfits",
+    "setBonuses",
+    "statusEffects",
+    "race",
+    "colours",
+    "combatMove",
+    "dialogue",
+    "encounters",
+    "sex",
+    "maps",
+    "txt",
 )
+
 subfolders_of = {
     "items": ("weapons", "clothing", "tattoos", "items", "patterns"),
     "race": ("bodyParts", "coveringTypes", "subspecies"),
+    "sex": ("managers", "actions"),
 }
 
 
@@ -114,7 +127,9 @@ def _bad_directory_structure(path: pathlib.Path):
 
 @lru_cache(maxsize=None)
 def _bad_suffix(suffix):
-    return bool(suffix not in (".xml", ".svg"))
+    # FIXME: PNGs are only warranted within maps subfolders. We don't support per folder
+    # suffixes though.
+    return bool(suffix not in (".xml", ".svg", ".png"))
 
 
 def file_status(file: bucket.FileMetadata) -> FileState: